new config with azure
